use int64 instead if int for both arguments and results

+++ client.go
@@ -6,8 +6,8 @@ type Client interface {
 	Get(key string) (string, error)
 	MGet(keys []string) ([]string, error)
 	Set(key string, value interface{}) error
-	SetEx(key string, expire int, value interface{}) error
-	Expire(key string, seconds int) error
+	SetEx(key string, expire int64, value interface{}) error
+	Expire(key string, seconds int64) error
 	Del(key string) error
 
 	LPush(key string, value string) error
@@ -18,6 +18,6 @@ type Client interface {
 	Incr(key string) error
 	IncrBy(key string, inc interface{}) (interface{}, error)
 
-	ZAdd(key string, score float64, value interface{}) (int, error)
-	ZCount(key string, min interface{}, max interface{}) (int, error)
+	ZAdd(key string, score float64, value interface{}) (int64, error)
+	ZCount(key string, min interface{}, max interface{}) (int64, error)
 }
